Add barebones Node struct
Basically does nothing as of now, just has some "persistent" state
(which is not really persistent as of now) and an election timer which
samples from a uniform distribution which is also hardcoded as of now

Current Log is also just a vector that is part of the persistent storage
struct, this should ideally be dealt with separately since it will
mostly just be appended to. For other persistent things just serializing
with Serde into a file should be enough.