Add raft log implementation
Each entry has 3 variables that are persisted in the state file: index,
term, length (of log message). After length, the actual log is written
in the file. However the actual log is never stored in memory, point
being that we can read it from the file when we need to apply it to our
state machine. That operation happens only once, so reading from disk
shouldn't be a big issue